requirements.txt
greenlet==3.1.1
numpy==2.1.1
pandas==2.2.3
prompt-toolkit==3.0.36
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
questionary==2.0.1
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
